requirements.txt
python-telegram-bot==20.7
anthropic==0.25.0
google-auth==2.29.0
google-api-python-client==2.125.0
google-auth-httplib2==0.2.0
gspread==6.1.2
